Refactor step definitions.
